[CIDMDEV-5132] - Migration Support - Read-Only Endpoint Templates/CloudBaseUrl
- Protected the following services by flag 'feature.migration.read.only.mode.enabled'
(returns 503 when read-only enabled) :

1. POST v1.1/baseURLs
2. POST v2.0/OS-KSCATALOG/endpointTemplates
3. PUT v2.0/OS-KSCATALOG/endpointTemplates/{endpointTemplateId}
4. DELETE v2.0/OS-KSCATALOG/endpointTemplates/{endpointTemplateId}

The v1.1 service to add a base URL (POST v1.1/baseURLs) is already feature flagged to disable (returning a 404) when the static property feature.v11.add.base.url.exposed=false. The migration read-only flag will therefore only affect this service when feature.v11.add.base.url.exposed=true.

+++ src/main/java/com/rackspace/idm/api/resource/cloud/v20/Cloud20VersionResource.java
@@ -801,7 +801,11 @@ public class Cloud20VersionResource {
             @Context HttpHeaders httpHeaders,
             @Context UriInfo uriInfo,
             @HeaderParam(X_AUTH_TOKEN) String authToken, EndpointTemplate endpoint) {
-        return cloud20Service.addEndpointTemplate(httpHeaders, uriInfo, authToken, endpoint).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.addEndpointTemplate(httpHeaders, uriInfo, authToken, endpoint).build();
+        }
     }
 
     @PUT
@@ -811,7 +815,11 @@ public class Cloud20VersionResource {
             @Context UriInfo uriInfo,
             @PathParam("endpointTemplateId") String endpointTemplateId,
             @HeaderParam(X_AUTH_TOKEN) String authToken, EndpointTemplate endpoint) {
-        return cloud20Service.updateEndpointTemplate(httpHeaders, uriInfo, authToken, endpointTemplateId, endpoint).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.updateEndpointTemplate(httpHeaders, uriInfo, authToken, endpointTemplateId, endpoint).build();
+        }
     }
 
     @GET
@@ -829,7 +837,11 @@ public class Cloud20VersionResource {
             @Context HttpHeaders httpHeaders,
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("endpointTemplateId") String enpdointTemplateId) {
-        return cloud20Service.deleteEndpointTemplate(httpHeaders, authToken, enpdointTemplateId).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.deleteEndpointTemplate(httpHeaders, authToken, enpdointTemplateId).build();
+        }
     }
 
     @GET
